rename normalization module
